Adding new holding checks for existing assets.
If the asset name matches an existing asset, that asset will be used for
the holding. Otherwise, a new asset will be created. If an existing
asset is used, the price of that asset will be updated to reflect the
price given in the add command (-p option). The change will be reflected
across all holding that contain that specific asset.